+++ contracts/netmap/contract.go
@@ -130,6 +130,11 @@ func _deploy(data any, isUpdate bool) {
 			storage.Delete(ctx, snapshotBlockKey)
 		}
 
+		if version < 24_000 {
+			// For whatever reason this was also stored here, not just in neofs contract.
+			const candidateFeeConfigKey = "InnerRingCandidateFee"
+			storage.Delete(ctx, append(configPrefix, []byte(candidateFeeConfigKey)...))
+		}
 		return
 	}